+++ app/imports/api/blocklist/BlocklistMethods.js
@@ -1,11 +1,18 @@
 import { Meteor } from 'meteor/meteor';
 import { BlocklistCollection } from './BlocklistCollection';
+import { IpAddressCollection } from '../ipaddress/IpAddressCollection';
 import 'meteor/dburles:collection-helpers';
 
 
 Meteor.methods({
     'blocklists.count': function () {
         return BlocklistCollection.find().count();
-    }
+    },
+    'blocklists.remove': function (blocklist_id) {
+        check(blocklist_id, String);
+
+        IpAddressCollection.remove({ blocklist_id: blocklist_id });
+        BlocklistCollection.remove(blocklist_id);
+    },
 });