+++ app/imports/api/blocklist/index.js
@@ -1,4 +1,5 @@
 import './BlocklistCollection';
 import './BlocklistSchema';
 import './BlocklistTable';
+import './BlocklistMethods';