+++ app/imports/ui/pages/edit-ipaddress-page.js
@@ -13,6 +13,10 @@ AutoForm.hooks({
   },
 });
 
+Template.Edit_IpAddress_Page.onCreated(function () {
+  this.subscribe('ipAddresses');
+});
+
 Template.Edit_IpAddress_Page.helpers({
   getDoc() {
     return IpAddressCollection.findOne(FlowRouter.getParam('_ipAddress_id'));
@@ -24,4 +28,18 @@ Template.Edit_IpAddress_Page.events({
   'click .cancel': function () {
     history.back();
   },
+  'click .delete': function (event, template) {
+    const ipAddress_id = FlowRouter.getParam('_ipAddress_id')
+    const ipAddress = IpAddressCollection.findOne(ipAddress_id);
+    if (ipAddress) {
+      if (confirm('Really delete "' + ipAddress.address + '"?')) {
+        Meteor.call('ipAddresses.remove', ipAddress._id, function (err, res) {
+          if (err)
+            alert(err);
+          else
+            history.back();
+        });
+      }
+    }
+  }
 });