+++ app/imports/ui/pages/list-blocklist-page.js
@@ -3,7 +3,7 @@ import { Template } from 'meteor/templating';
 import { ReactiveVar } from 'meteor/reactive-var';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
-Template.List_Blocklist_Page.onCreated(function() {
+Template.List_Blocklist_Page.onCreated(function () {
   this.subscribe('ipAddresses');
   this.blocklistCount = new ReactiveVar();
 
@@ -13,7 +13,7 @@ Template.List_Blocklist_Page.onCreated(function() {
 });
 
 Template.List_Blocklist_Page.helpers({
-  blocklistCount: function() {
+  blocklistCount: function () {
     return Template.instance().blocklistCount.get();
   }
 });
@@ -22,27 +22,6 @@ Template.List_Blocklist_Page.helpers({
 Template.List_Blocklist_Actions_Cell.helpers({
   canShow: function () {
     return !!Meteor.user();
-  },
-  onError: function () {
-    return function (error) {
-      console.log('Error', error);
-    };
-  },
-  onSuccess: function () {
-    return function (result) {
-      console.log('Deleted');
-    };
-  },
-  beforeRemove: function () {
-    return function (collection, id) {
-      var doc = collection.findOne(id);
-      if (confirm('Really delete "' + doc.name + '"?')) {
-        const self = this;
-        Meteor.call('ipAddresses.removeFromBlocklist', doc._id, function (err, res) {
-          if (!err) self.remove();
-        });
-      }
-    };
   }
 });
 
@@ -52,6 +31,14 @@ Template.List_Blocklist_Actions_Cell.events({
   },
   'click .edit': function () {
     FlowRouter.go('Edit_Blocklist_Page', { _blocklist_id: this._id });
+  },
+  'click .delete': function (event, template) {
+    if (confirm('Really delete "' + this.name + '"?')) {
+      Meteor.call('blocklists.remove', this._id, function (err, res) {
+        if (err)
+          alert(err);
+      });
+    }
   }
 });