+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Version 0.1.1
 
 - Added Version information
+- Added blocklists.remove method
+- Added ipAddresses.remove method
+- Removed aldeed:delete-button (quickRemoveButton) use regular buttons
+- Added delete buttons to edit pages
 
 ## Version 0.1.0 (Initial release)